challenges/web/jwthe/Solve/requirements.txt
gmpy2
pyasn1
pycryptodome
pyjwt
fastecdsa